# skip missing lookup if nothing is missing in CPU hist partition kernel.