--- a/tests/cpp/tree/test_quantile_hist.cc
+++ b/tests/cpp/tree/test_quantile_hist.cc
@@ -251,14 +251,14 @@ class QuantileHistMock : public QuantileHistMaker {
     }
 
     void TestApplySplit(const GHistIndexBlockMatrix& quantile_index_block,
-                        RegTree& tree) {
+                        const RegTree& tree) {
       std::vector<GradientPair> row_gpairs =
           { {1.23f, 0.24f}, {0.24f, 0.25f}, {0.26f, 0.27f}, {2.27f, 0.28f},
             {0.27f, 0.29f}, {0.37f, 0.39f}, {-0.47f, 0.49f}, {0.57f, 0.59f} };
       size_t constexpr kMaxBins = 4;
 
       // try out different sparsity to get different number of missing values
-      for (double sparsity = 0.0; sparsity < 0.25; sparsity += 0.1) {
+      for (double sparsity : {0.0, 0.1, 0.2}) {
         // kNRows samples with kNCols features
         auto dmat = RandomDataGenerator(kNRows, kNCols, sparsity).Seed(3).GenerateDMatrix();