While looking at the profile, I noticed that PartitionDenseKernel was taking surprising %% of time;
Looking at it a bit, large part of the cost comes from the column.IsMissing check; This check:
1) does lookup in `vector<bool>`
2) adds one more condition to the hot loop
3) fills the caches with the data from that vector<bool>

In my use-cases, as well as in several publicly available datasets (like
higgs), nothing is missing, so we can use that information and create a
separate path which won't do this check at all.

Benchmarking on higgs dataset (10m samples, 28 features, 100 trees, 256 leaves, 256 bins) on 4 core macbook pro (Kaby Lake Core i5) shows ~10% improvement in training time.
I do not have other machine to test it on, so I cannot really claim how much of the win will hold on other systems/other datasets.

My results by running 10 runs of old/new:
Old:
```
[21:54:18] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 18.651 sec in all
[21:55:00] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 18.7015 sec in all
[21:55:42] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 18.6216 sec in all
[21:56:24] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 19.4026 sec in all
[21:57:05] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 18.0353 sec in all
[21:57:45] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 18.0898 sec in all
[21:58:26] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 18.457 sec in all
[21:59:07] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 18.1343 sec in all
[21:59:49] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 18.2275 sec in all
[22:00:30] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 18.3709 sec in all
```

**Avg = 18.46s**

New:

```
[21:53:56] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 16.3706 sec in all
[21:54:38] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 17.1902 sec in all
[21:55:20] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 17.0527 sec in all
[21:56:01] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 16.6894 sec in all
[21:56:44] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 16.6933 sec in all
[21:57:24] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 16.8017 sec in all
[21:58:05] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 16.6709 sec in all
[21:58:46] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 16.9324 sec in all
[21:59:27] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 16.9194 sec in all
[22:00:08] INFO: /Users/oleksandr/projects/xgboost/src/cli_main.cc:283: update end, 16.5891 sec in all
```

**Avg = 16.79s** 

A few further improvements in the same area to consider doing later:
1) make it per column, so that one column with missing values won't cause IsMissing calls for all other columns
2) do not even store that vector<bool> if it's unused; While not terribly huge, it will be still ~(samples * features / 8) bytes;