@SmirnovEgorRu 2020-05-07
PartitionDenseKernel is top2 hotspot after building hist for most of data sets now (they are even on par in some cases), so for me it's not surprising.
Overall, good to see performance improvement :)