@okuvshynov 2020-05-07
@trivialfis - sure, will add.
@SmirnovEgorRu - very interesting! Do you have some aggregated profile data, taken on different hardware with different datasets?