@SmirnovEgorRu 2020-05-07
@okuvshynov, I have this picture, it shows % of total time for each XGB kernel for a few data sets (obtained on 2 socket 24 cores Xeon). Probably, it's massive, but I hope it will be anyway helpful for you.

![image](https://user-images.githubusercontent.com/18555970/81310738-2db75700-908d-11ea-82e4-5a838d1351b7.png)