@ShvetsKS 2020-05-07
@okuvshynov 
About further optimizations. We could allocate only O(n_features) additional space as a flags for each column ('true' means - missing values are exist in certain column).
Seems to do it we could find out which features were not presented in certain instance
https://github.com/dmlc/xgboost/pull/5644/files#diff-148a32528d3b279f484421bd485ff01fR258
and add 'true' for it.