@okuvshynov 2020-05-08
@SmirnovEgorRu - thank you! To confirm, this picture is showing wall time spent, not CPU time/cycles, right? Do you know the details about SyncHistogram cost, what's the breakdown between:
1) not being able to utilize all the cores for that operation
2) synchronization overhead, if any
3) just doing more work compared to low-core scenario, because there are more threads with more histograms to merge?

My 4 core machine doesn't have the same pattern, obviously. 
Also, what is the best place to discuss this in more details, so we don't flood this pull request?